# api.yaml
swagger: "2.0"
info:
  description: SAM's API Swagger documentation
  version: 0.1
# Some services need a JWT authorization token to be parsed in the header (Authorization)
securityDefinitions:  
  JWT:  
    type: apiKey  
    in: header
    name: Authorization
tags:
- name: "User"
  description: "Operations about user"
paths:
# Authenticate a User
  /user/login:
    post:
      tags:
      - "User"
      summary: "Login user into the system"
      description: ""
      operationId: "authenticate"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "application/json"
      parameters:
      - in: "formData"
        name: "email"
        schema:
            type: string
        description: "Email of the user to be authenticated"
        required: true
      - in: "formData"
        name: "psw"
        schema:
            type: string
        description: "Password of the user to be authenticated"
        required: true
      responses:
        '200':
            description: Returns an authentication token
            examples:
              # Returns a JSON object as the following example:
              application/json: 
                { 
                  "/user/login":{
                    "ID": 1,
                    "avatar": "path/to/avatar/ID",
                    "status": 200,
                    "token": (...) eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9 (...)
                  }
                }
        '400':
            description: The email or password cannot be empty
        '401':
            description: Authentication failure
        '404':
            description: A user with the specified email was not found.
        '5XX':
            description: Unexpected error.
  # Add a new User
  /user:
    post:
      tags:
      - "User"
      summary: "Create user"
      description: ""
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User object that needs to be added"
        required: true
        example:
            {
              'g-recaptcha-response': 'secret_recaptcha_response',
              'email': 'new@user.com',
              'firstName': 'Test',
              'lastName': 'User',
              'psw': 'Super_Secret_Plain_Password',
              'avatar': 'path/to/avatar/img'
            }
      responses:
        '200':
            description: The user was created 
            examples:
              # Returns a JSON object as the following example:
              application/json: 
                { 
                  "/user":{
                    "status": 200,
                  }
                }
        '400':
            description: Invalid parsed JSON object
        '5XX':
            description: Unexpected error.
  # Get a User
  /user/{email}:
    get:
      tags:
      - "User"
      security:  
      - JWT: []
      summary: "Find User by email"
      description: "Returns a single user"
      operationId: "getUser"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "path"
        description: "Email of the user to return"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Invalid email supplied"
        403:
          description: "Authentication failure"
        404:
          description: "User not found"
    put:
      tags:
      - "User"
      security:  
      - JWT: []
      summary: "Update User"
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "path"
        description: "Email of the user to update"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        example:
            {
              'email': 'new.updated@user.com',
              'psw': 'Super_Secret_Plain_Password_Updated_or_Same',
              'firstName': 'Test Updated',
              'lastName': 'User Updated',
              'avatar': 'path/to/avatar/img_updated'
            }
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Invalid email supplied"
        403:
          description: "Authentication failure"
        404:
          description: "User not found"
    delete:
      tags:
      - "User"
      security:  
      - JWT: []
      summary: "Delete a User"
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "path"
        description: "Email of the user to delete"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Invalid email supplied"
        403:
          description: "Authentication failure"
        404:
          description: "User not found"